regulation_task/pyproject.toml
[project]
name = "regulation-task"
version = "0.1.0"
description = "Regulatory Compliance Document Processor"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.49.0",
    "chromadb>=0.6.3",
    "fastapi>=0.115.11",
    "langchain>=0.3.19",
    "langchain-chroma>=0.2.2",
    "langchain-openai>=0.3.7",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.2.1",
    "pymupdf>=1.25.3",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.0",
    "jinja2>=3.1.3",
]

[project.optional-dependencies]
dev = [
    "black>=24.2.0",
    "isort>=5.13.2",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100
